Add graph basics: BFS, DFS, components, and notes
Introduced foundational graph algorithms and concepts including BFS and DFS traversals, connected components (BFS/DFS), and various graph representations in Java. Added comprehensive notes covering graph introduction, traversal techniques, connected components, and representation, along with example problems and solutions. Also updated RecoverBST.java in the Trees section.

+++ Problems/Trees/BinarySearchTrees/Medium/RecoverBST.java
@@ -83,8 +83,8 @@ public class RecoverBST {
             root = root.right;
         }
 
-        int temp = first.data;
-        first.data = second.data;
+        int temp = (Integer) first.data;
+        first.data = (Integer) second.data;
         second.data = temp;
     }
